--- internal/appsec/emitter/graphqlsec/init.go
@@ -1,15 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016 Datadog, Inc.
-
-package graphqlsec
-
-import (
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec"
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec/listener/graphqlsec"
-)
-
-func init() {
-	appsec.AddWAFEventListener(graphqlsec.Install)
-}
